Make `Style.with_reset` more explicit as `prefix_with_reset`
As the behavior is probably a bit unclear and everybody constructing or
destructuring it by hand will now have to take this field into account
we should be as explicit as possible. @fdncred 's suggestion was
`prefix_with_reset`.
I also tried to be even more detailed in the fields docstring for
anybody interacting it after `0.47.0`. @alexkunde check me if I am wrong
here.